+++ Realization/backend/vs/src/main/java/eu/bcvsolutions/idm/vs/bulk/action/impl/VsRequestCancelBulkAction.java
@@ -1,6 +1,15 @@
 package eu.bcvsolutions.idm.vs.bulk.action.impl;
 
+import java.util.List;
+
+import org.apache.logging.log4j.util.Strings;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Description;
+import org.springframework.stereotype.Component;
+import org.springframework.util.Assert;
+
 import com.google.common.collect.Lists;
+
 import eu.bcvsolutions.idm.core.api.bulk.action.AbstractBulkAction;
 import eu.bcvsolutions.idm.core.api.domain.OperationState;
 import eu.bcvsolutions.idm.core.api.entity.OperationResult;
@@ -15,17 +24,6 @@ import eu.bcvsolutions.idm.vs.domain.VirtualSystemGroupPermission;
 import eu.bcvsolutions.idm.vs.dto.VsRequestDto;
 import eu.bcvsolutions.idm.vs.dto.filter.VsRequestFilter;
 import eu.bcvsolutions.idm.vs.service.api.VsRequestService;
-import java.util.Collection;
-import java.util.List;
-import java.util.UUID;
-import javax.persistence.EntityManager;
-import org.apache.logging.log4j.util.Strings;
-import org.hibernate.Session;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Description;
-import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.util.Assert;
 
 /**
  * Bulk operation for cancel virtual system's requests.